refactor(image-editor): extract the stroke engine into a DOM-free paint core
The brush/pencil/eraser engine lived in web/ and could only run in a browser,
so nothing headless could paint a pixel. It is pure Canvas2D apart from two
window.document.createElement("canvas") calls for the brush stamp cache.

Move it to @nodetool-ai/image-editor/painting.js behind an injectable surface
factory (setPaintSurfaceFactory), defaulting to the DOM so the web app is
unchanged. The stamp cache is typed structurally, so both HTMLCanvasElement
and @napi-rs/canvas satisfy it with no cast.

web/ keeps every existing import path as a re-export: there is now exactly one
definition of the engine and of the stroke settings types.
