+++ packages/image-editor/README.md
@@ -1,8 +1,8 @@
 # @nodetool-ai/image-editor
 
-Shared image-editor types, dependency hashing, and seeded layer templates for [NodeTool](https://nodetool.ai).
+Shared image-editor types, dependency hashing, seeded layer templates, and the paint core for [NodeTool](https://nodetool.ai).
 
-The pure type layer behind NodeTool's layered sketch editor: the persisted document shape, per-layer generation bindings (workflow-bound and direct text-to-image / image-to-image / inpaint), version history, and the content hash that detects when a layer is stale. Kept free of the web editor implementation so both the browser and server can depend on it.
+The pure layer behind NodeTool's layered sketch editor: the persisted document shape, per-layer generation bindings (workflow-bound and direct text-to-image / image-to-image / inpaint), version history, the content hash that detects when a layer is stale, and the brush/pencil/eraser stroke engine. Kept free of the web editor implementation so both the browser and server can depend on it.
 
 ## Install
 
@@ -32,6 +32,44 @@ The Node-only `computeDependencyHash` (and its `DependencyHashInput`) is not re-
 ```ts
 import { computeDependencyHash } from "@nodetool-ai/image-editor/dependencyHash";
 ```
+
+## Paint core
+
+`@nodetool-ai/image-editor/painting.js` is the brush/pencil/eraser stroke
+engine — pressure dynamics, dab spacing, brush-stamp caching, supersampling,
+dirty-rect tracking — plus the settings shapes it reads (`BrushSettings`,
+`PencilSettings`, `EraserSettings`, `DEFAULT_BRUSH_SETTINGS`, …). The web
+sketch editor re-exports it; there is no second copy.
+
+It is plain Canvas2D with one seam. The engine allocates the off-screen bitmaps
+for its stamp cache through `createPaintSurface`, which defaults to
+`window.document.createElement("canvas")`. A headless host swaps that out:
+
+```ts
+import { createCanvas } from "@napi-rs/canvas";
+import {
+  setPaintSurfaceFactory,
+  drawBrushStroke,
+  DEFAULT_BRUSH_SETTINGS
+} from "@nodetool-ai/image-editor/painting.js";
+
+setPaintSurfaceFactory(createCanvas);
+
+const canvas = createCanvas(256, 256);
+drawBrushStroke(
+  { x: 32, y: 48 },
+  { x: 224, y: 176 },
+  { ...DEFAULT_BRUSH_SETTINGS, size: 24, color: "#ff2d55" },
+  canvas.getContext("2d"),
+  undefined,
+  { current: null },
+  new Map()
+);
+```
+
+`setPaintSurfaceFactory(null)` restores the browser default. `PaintContext2D`
+and `PaintSurface` are structural types, so a DOM `CanvasRenderingContext2D`
+and a skia canvas both satisfy them without a cast.
 
 ## Usage
