+++ packages/image-editor/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@nodetool-ai/image-editor",
   "version": "0.7.0-rc.34",
-  "description": "Shared image-editor types, dependency hashing, and seeded layer templates",
+  "description": "Shared image-editor types, dependency hashing, seeded layer templates, and the host-neutral brush/pencil/eraser paint core",
   "type": "module",
   "engines": {
     "node": ">=22.0.0 <23.0.0"
@@ -31,6 +31,12 @@
       "types": "./dist/dependencyHash.d.ts",
       "import": "./dist/dependencyHash.js",
       "default": "./dist/dependencyHash.js"
+    },
+    "./painting.js": {
+      "nodetool-dev": "./src/painting/index.ts",
+      "types": "./dist/painting/index.d.ts",
+      "import": "./dist/painting/index.js",
+      "default": "./dist/painting/index.js"
     }
   },
   "license": "MIT",
