+++ web/src/components/sketch/types/tools.ts
@@ -3,9 +3,51 @@
  *
  * Tool union, per-tool settings interfaces, default constants, pressure/assist
  * configuration, segmentation types, and tool-classification utilities.
+ *
+ * Everything the headless stroke engine reads — brush type, pen pressure,
+ * stroke assist, and the brush/pencil/eraser settings with their defaults —
+ * lives in the paint core (`@nodetool-ai/image-editor/painting.js`) and is
+ * re-exported here so the sketch editor's own imports are unchanged.
  */
 
 import type { SelectSettings } from "./selection";
+import {
+  DEFAULT_BRUSH_SETTINGS,
+  DEFAULT_ERASER_SETTINGS,
+  DEFAULT_PENCIL_SETTINGS,
+  DEFAULT_PEN_PRESSURE
+} from "@nodetool-ai/image-editor/painting.js";
+import type {
+  BrushSettings,
+  EraserSettings,
+  PencilSettings,
+  PenPressureSettings
+} from "@nodetool-ai/image-editor/painting.js";
+
+export type {
+  BrushType,
+  BrushSettings,
+  PencilSettings,
+  EraserMode,
+  EraserSettings,
+  PenPressureSettings,
+  StrokeAssistMode,
+  StrokeAssistSnapMode,
+  StrokeAssistPreset,
+  StrokeAssistSettings
+} from "@nodetool-ai/image-editor/painting.js";
+export {
+  DEFAULT_PRESSURE_MIN_SCALE,
+  DEFAULT_PRESSURE_CURVE,
+  DEFAULT_PEN_PRESSURE,
+  DEFAULT_BRUSH_SETTINGS,
+  DEFAULT_PENCIL_SETTINGS,
+  DEFAULT_ERASER_SETTINGS,
+  createStrokeAssistPreset,
+  resolveStrokeAssistSettings,
+  mergePenPressureIntoBrush,
+  mergePenPressureIntoPencil
+} from "@nodetool-ai/image-editor/painting.js";
 
 // ─── Tool Types ───────────────────────────────────────────────────────────────
 
@@ -28,21 +70,7 @@ export type SketchTool =
 
 export type ShapeToolType = "line" | "rectangle" | "ellipse" | "arrow";
 
-export type BrushType = "round" | "soft" | "airbrush" | "spray";
-
 // ─── Pen Pressure ─────────────────────────────────────────────────────────────
-
-/**
- * Default light-press scale (6% of full size at minimum pressure; see `strokePressureMultiplier` in drawingUtils).
- * Lower = thinner light strokes and a wider thin→thick range (also adjustable per tool in settings).
- */
-export const DEFAULT_PRESSURE_MIN_SCALE = 0.06;
-
-/**
- * Raw pen pressure is raised to this power before mapping to `[pressureMinScale, 1]`.
- * `1` = linear; values above 1 need firmer pressure to reach full size (more contrast in mid/high pressure).
- */
-export const DEFAULT_PRESSURE_CURVE = 1;
 
 /** Stored `pressureMinScale` range surfaced by the Light end control (see eased slider mapping). */
 const PRESSURE_MIN_SCALE_UI_MIN = 0.02;
@@ -71,237 +99,7 @@ export function pressureMinScaleToSliderUnit(m: number): number {
   return 1 - Math.sqrt(Math.max(0, Math.min(1, 1 - r)));
 }
 
-/**
- * Global pen/tablet pressure (single source of truth for drawing).
- * {@link BrushSettings} / {@link PencilSettings} still include the same fields for future per-tool expansion.
- */
-export interface PenPressureSettings {
-  pressureSensitivity: boolean;
-  pressureAffects: "size" | "opacity" | "both";
-  pressureMinScale: number;
-  pressureCurve: number;
-}
-
-export const DEFAULT_PEN_PRESSURE: PenPressureSettings = {
-  pressureSensitivity: true,
-  pressureAffects: "both",
-  pressureMinScale: DEFAULT_PRESSURE_MIN_SCALE,
-  pressureCurve: DEFAULT_PRESSURE_CURVE
-};
-
-// ─── Stroke Assist ────────────────────────────────────────────────────────────
-
-export type StrokeAssistMode = "stabilizer" | "lazy";
-
-export type StrokeAssistSnapMode = "off" | "angle";
-
-export type StrokeAssistPreset = "smooth" | "lazy" | "inking" | "custom";
-
-export interface StrokeAssistSettings {
-  preset: StrokeAssistPreset;
-  mode: StrokeAssistMode;
-  /** Main assist amount: 0 = off, 1 = maximum effect. */
-  strength: number;
-  snapMode: StrokeAssistSnapMode;
-  /** Blend toward the snapped guide when snapping is enabled. */
-  snapStrength: number;
-  /** Angle step in degrees for angle snap. */
-  angleIncrement: number;
-}
-
-const DEFAULT_STROKE_ASSIST_SETTINGS: StrokeAssistSettings = {
-  preset: "custom",
-  mode: "stabilizer",
-  strength: 0,
-  snapMode: "off",
-  snapStrength: 0.75,
-  angleIncrement: 45
-};
-
-export function createStrokeAssistPreset(
-  preset: Exclude<StrokeAssistPreset, "custom">
-): StrokeAssistSettings {
-  switch (preset) {
-    case "smooth":
-      return {
-        preset,
-        mode: "stabilizer",
-        strength: 0.65,
-        snapMode: "off",
-        snapStrength: DEFAULT_STROKE_ASSIST_SETTINGS.snapStrength,
-        angleIncrement: 45
-      };
-    case "lazy":
-      return {
-        preset,
-        mode: "lazy",
-        strength: 0.6,
-        snapMode: "off",
-        snapStrength: DEFAULT_STROKE_ASSIST_SETTINGS.snapStrength,
-        angleIncrement: 45
-      };
-    case "inking":
-      return {
-        preset,
-        mode: "lazy",
-        strength: 0.45,
-        snapMode: "angle",
-        snapStrength: 0.9,
-        angleIncrement: 45
-      };
-  }
-}
-
-const STROKE_ASSIST_ANGLE_INCREMENTS = [15, 30, 45, 90] as const;
-
-function normalizedStrokeAssistAngleIncrement(value: unknown, fallback: number): number {
-  if (typeof value !== "number" || !Number.isFinite(value)) {
-    return fallback;
-  }
-  const nearest = STROKE_ASSIST_ANGLE_INCREMENTS.reduce((best, candidate) =>
-    Math.abs(candidate - value) < Math.abs(best - value) ? candidate : best
-  );
-  return nearest;
-}
-
-function normalizedUnitScalar(value: unknown, fallback: number): number {
-  if (typeof value !== "number" || !Number.isFinite(value)) {
-    return fallback;
-  }
-  return Math.min(1, Math.max(0, value));
-}
-
-export function resolveStrokeAssistSettings(
-  legacyStabilizer: number | undefined,
-  strokeAssist: Partial<StrokeAssistSettings> | undefined
-): StrokeAssistSettings {
-  const presetBase =
-    strokeAssist?.preset && strokeAssist.preset !== "custom"
-      ? createStrokeAssistPreset(strokeAssist.preset)
-      : DEFAULT_STROKE_ASSIST_SETTINGS;
-  const merged = {
-    ...presetBase,
-    ...(strokeAssist ?? {})
-  };
-  if (!strokeAssist) {
-    merged.strength = normalizedUnitScalar(
-      legacyStabilizer,
-      DEFAULT_STROKE_ASSIST_SETTINGS.strength
-    );
-    merged.mode = "stabilizer";
-    merged.snapMode = "off";
-    merged.preset = "custom";
-  }
-  const shouldPreferLegacyStabilizer =
-    typeof legacyStabilizer === "number" &&
-    legacyStabilizer > 0 &&
-    merged.preset === "custom" &&
-    merged.mode === "stabilizer" &&
-    merged.strength === DEFAULT_STROKE_ASSIST_SETTINGS.strength &&
-    merged.snapMode === DEFAULT_STROKE_ASSIST_SETTINGS.snapMode;
-  if (shouldPreferLegacyStabilizer) {
-    merged.strength = legacyStabilizer;
-  }
-  return {
-    preset: merged.preset ?? "custom",
-    mode: merged.mode === "lazy" ? "lazy" : "stabilizer",
-    strength: normalizedUnitScalar(
-      merged.strength,
-      DEFAULT_STROKE_ASSIST_SETTINGS.strength
-    ),
-    snapMode: merged.snapMode === "angle" ? "angle" : "off",
-    snapStrength: normalizedUnitScalar(
-      merged.snapStrength,
-      DEFAULT_STROKE_ASSIST_SETTINGS.snapStrength
-    ),
-    angleIncrement: normalizedStrokeAssistAngleIncrement(
-      merged.angleIncrement,
-      DEFAULT_STROKE_ASSIST_SETTINGS.angleIncrement
-    )
-  };
-}
-
 // ─── Tool Settings Interfaces ─────────────────────────────────────────────────
-
-export interface BrushSettings {
-  size: number;
-  opacity: number;
-  hardness: number;
-  color: string;
-  brushType: BrushType;
-  pressureSensitivity: boolean;
-  pressureAffects: "size" | "opacity" | "both";
-  /** Light-press scale (typically 0.02–0.5). Defaults: {@link DEFAULT_PRESSURE_MIN_SCALE}. */
-  pressureMinScale: number;
-  /** Pressure response curve exponent (typically 0.5–2.5). Defaults: {@link DEFAULT_PRESSURE_CURVE}. */
-  pressureCurve: number;
-  roundness: number; // 0.1 to 1.0 (1.0 = perfect circle)
-  angle: number; // 0 to 360 degrees
-  /** Stroke stabilizer strength: 0 = off, 1 = maximum smoothing. */
-  stabilizer: number;
-  /** New stroke input assist model. Falls back to legacy `stabilizer` when absent. */
-  strokeAssist?: StrokeAssistSettings;
-}
-
-export interface PencilSettings {
-  size: number;
-  opacity: number;
-  color: string;
-  pressureSensitivity: boolean;
-  pressureAffects: "size" | "opacity" | "both";
-  /** @see {@link BrushSettings.pressureMinScale} */
-  pressureMinScale: number;
-  /** @see {@link BrushSettings.pressureCurve} */
-  pressureCurve: number;
-  /** Stroke stabilizer strength: 0 = off, 1 = maximum smoothing. */
-  stabilizer: number;
-  /** New stroke input assist model. Falls back to legacy `stabilizer` when absent. */
-  strokeAssist?: StrokeAssistSettings;
-  /**
-   * When true, dabs are drawn as crisp N×N filled squares centered on the
-   * pixel under the cursor (pixel-art style) instead of antialiased circles.
-   * Optional for back-compat with older serialized documents; treat absent
-   * as `true` (the default for new pencils).
-   */
-  pixelPerfect?: boolean;
-}
-
-/** Apply global {@link ToolSettings.penPressure} for paint engines (brush/pencil store strips these for UI). */
-export function mergePenPressureIntoBrush(
-  brush: BrushSettings,
-  penPressure: PenPressureSettings | undefined
-): BrushSettings {
-  return {
-    ...brush,
-    ...DEFAULT_PEN_PRESSURE,
-    ...(penPressure ?? {})
-  };
-}
-
-/** Apply global {@link ToolSettings.penPressure} for paint engines (brush/pencil store strips these for UI). */
-export function mergePenPressureIntoPencil(
-  pencil: PencilSettings,
-  penPressure: PenPressureSettings | undefined
-): PencilSettings {
-  return {
-    ...pencil,
-    ...DEFAULT_PEN_PRESSURE,
-    ...(penPressure ?? {})
-  };
-}
-
-/** Brush: same stamp as Brush tool (`drawBrushStroke`). Pencil: same as Pencil tool (`drawPencilStroke`). */
-export type EraserMode = "brush" | "pencil";
-
-export interface EraserSettings {
-  size: number;
-  opacity: number;
-  mode: EraserMode;
-  /** Stroke stabilizer strength: 0 = off, 1 = maximum smoothing. */
-  stabilizer: number;
-  /** New stroke input assist model. Falls back to legacy `stabilizer` when absent. */
-  strokeAssist?: StrokeAssistSettings;
-}
 
 export interface ShapeSettings {
   shapeType: ShapeToolType;
@@ -574,43 +372,6 @@ export interface ToolSettings {
 }
 
 // ─── Defaults ─────────────────────────────────────────────────────────────────
-
-export const DEFAULT_BRUSH_SETTINGS: BrushSettings = {
-  size: 8,
-  opacity: 1,
-  hardness: 0.65,
-  color: "#ffffff",
-  brushType: "round",
-  pressureSensitivity: true,
-  pressureAffects: "both",
-  pressureMinScale: DEFAULT_PRESSURE_MIN_SCALE,
-  pressureCurve: DEFAULT_PRESSURE_CURVE,
-  roundness: 1.0,
-  angle: 0,
-  stabilizer: 0,
-  strokeAssist: { ...DEFAULT_STROKE_ASSIST_SETTINGS }
-};
-
-export const DEFAULT_PENCIL_SETTINGS: PencilSettings = {
-  size: 2,
-  opacity: 1,
-  color: "#ffffff",
-  pressureSensitivity: true,
-  pressureAffects: "both",
-  pressureMinScale: DEFAULT_PRESSURE_MIN_SCALE,
-  pressureCurve: DEFAULT_PRESSURE_CURVE,
-  stabilizer: 0,
-  strokeAssist: { ...DEFAULT_STROKE_ASSIST_SETTINGS },
-  pixelPerfect: true
-};
-
-export const DEFAULT_ERASER_SETTINGS: EraserSettings = {
-  size: 14,
-  opacity: 1,
-  mode: "brush",
-  stabilizer: 0,
-  strokeAssist: { ...DEFAULT_STROKE_ASSIST_SETTINGS }
-};
 
 export const DEFAULT_SHAPE_SETTINGS: ShapeSettings = {
   shapeType: "rectangle",
